feat: Add Playwright E2E tests
This commit adds Playwright E2E tests for the application.

It includes: - Installation of Playwright and its dependencies. -
Configuration file for Playwright. - A basic test to verify the front
page loads correctly. - Configuration for the web server to run during
tests. - Ignoring the e2e tests in jest. - Configuring the github
workflow to run e2e tests and upload the test reports.